# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: C4ISR Operational Platform API
  version: 1.0.0
  license:
    name: Internal use
    identifier: LicenseRef-Internal
  description: |
    Versioned browser contract for the C2 and operational-awareness core.
    Observations and telemetry are evidence/ingestion boundaries; tracks and
    assets are current operational projections.
servers:
  - url: http://{host}:{port}
    description: Local development
    variables:
      host: { default: localhost }
      port: { default: '8080' }
...
paths:
...
  /api/v1/auth/me:
    get:
      tags: [identity]
      operationId: getCurrentOperator
      summary: Get current operator
      responses:
        '200':
          description: Authenticated operator and effective permissions
          content:
            application/json:
              schema:
                type: object
                required: [operator, permissions]
                properties:
                  operator: { $ref: '#/components/schemas/CurrentOperator' }
                  permissions:
                    type: array
                    items: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/realtime:
    get:
      tags: [system]
      security: [{ BearerAuth: [] }]
      description: WebSocket endpoint. Browser clients pass the bearer token as access_token during upgrade.
      operationId: realtime
      summary: Realtime
      responses:
        '101': { description: WebSocket upgrade }
        '200': { description: WebSocket negotiation response }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    CurrentOperator:
      type: object
      required: [id, name, role]
      properties: { id: { type: string }, name: { type: string }, role: { type: string } }
...
    Error:
      type: object
      required: [error]
      properties: { error: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string } } } }
...
  responses:
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Authorization denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
security:
  - BearerAuth: []